+++ src/components/modals/UpdateOrCreateEducationModal.vue
@@ -5,6 +5,7 @@
     import { useToast } from 'vue-toastification';
 
     import type { UserEducation } from '@type/entities';
+    import type { FormErrors } from "@type/responses"
 
     import { createEducation, updateEducation, deleteEducation } from '@services/educationService';
 
@@ -20,9 +21,12 @@
 
     const toast = useToast()
 
+    const isSubmitLoading = ref(false)
+    const isDeleteLoading = ref(false)
+
     const educations = defineModel<UserEducation[]>()
 
-    const errors = ref<any>()
+    const errors = ref<FormErrors>()
 
     const currentEducation = ref<UserEducation>({
         id: '',
@@ -34,10 +38,13 @@
     })
 
     const handleSubmit = async ()=>{
-        errors.value = null
+        if (isSubmitLoading.value) return
 
-        if(props.isEditing){
-            try {
+        try {
+            isSubmitLoading.value = true
+            errors.value = {}
+
+            if(props.isEditing){
                 const { id, userId, ...educationPayload } = currentEducation.value 
                 const res = await updateEducation(currentEducation.value.id, educationPayload)
                 
@@ -55,29 +62,29 @@
                 }
 
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
-            }
-        }else{
-            try {
+            }else{
                 const { id, userId, ...educationPayload } = currentEducation.value 
                 const res = await createEducation(educationPayload)
                 
                 educations.value?.push(res.data)
 
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
             }
+            emit('close')
+        } catch (error: any) {
+            errors.value = error.response?.data?.errors
+            toast.error(error.response?.data?.message)
+        }finally{
+            isSubmitLoading.value = false
         }
     }
 
     const handleDelete = async ()=>{
+        if (isDeleteLoading.value) return
+
         try {
+            isDeleteLoading.value = true
+
             const res = await deleteEducation(currentEducation.value.id)
 
             educations.value = educations.value?.filter(
@@ -88,6 +95,8 @@
             emit('close')
         } catch (error:any) {
             toast.error(error.response?.data?.message)
+        }finally{
+            isDeleteLoading.value = false
         }
     }
 
@@ -96,7 +105,7 @@
         ([isEditing, education]) => {
             if (isEditing &&  education) {
                 currentEducation.value = { ...education }
-                errors.value = ""
+                errors.value = {}
             }else {
                 currentEducation.value = {
                     id: '',
@@ -106,7 +115,7 @@
                     endDate:'',
                     userId: ''
                 }
-                errors.value = ""
+                errors.value = {}
             }
         },
         { immediate: true }
@@ -116,13 +125,17 @@
 <template>
     <BaseModal :open="open" @close="emit('close')">
         <h1 style="text-align: center;">{{ isEditing ? "Update Education" : "Create Education" }}</h1>
-        <Input label="School Name" placeholder="PNV" v-model="currentEducation.name"/>
-        <Input label="Degree" placeholder="Information Technology" v-model="currentEducation.degree"/>
-        <Input label="Start date" placeholder="11/11/2021" v-model="currentEducation.startDate" type="date"/>
-        <Input label="End date" placeholder="11/12/2021" v-model="currentEducation.endDate" type="date"/>
+        <Input :error="errors?.name?.[0]" label="School Name" placeholder="PNV" v-model="currentEducation.name"/>
+        <Input :error="errors?.degree?.[0]" label="Degree" placeholder="Information Technology" v-model="currentEducation.degree"/>
+        <Input :error="errors?.startDate?.[0]" label="Start date" placeholder="11/11/2021" v-model="currentEducation.startDate" type="date"/>
+        <Input :error="errors?.endDate?.[0]" label="End date" placeholder="11/12/2021" v-model="currentEducation.endDate" type="date"/>
         <div style="display: flex; gap:15px; align-items: center;">
-            <button class="btn" @click="handleSubmit">Submit</button>
-            <button v-if="isEditing" class="btn" @click="handleDelete">Delete</button>
+            <button class="btn" :disabled="isSubmitLoading" @click="handleSubmit">
+                {{ isSubmitLoading ? "Loading..." : "Submit" }}
+            </button>
+            <button v-if="isEditing" :disabled="isDeleteLoading" class="btn" @click="handleDelete">
+                {{ isDeleteLoading ? "Deleting..." : "Delete" }}
+            </button>
         </div>
     </BaseModal>
 </template>
